requirements.txt
pytesseract
Pillow
openpyxl
